Move CSV writer and add function to update rpeak threshold based on garment type
